# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@daytona.com
    name: Daytona Platforms Inc.
    url: https://www.daytona.io
  description: Daytona AI platform API Docs
  title: Daytona
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
...
  /sandbox/{sandboxId}/labels:
    put:
      operationId: replaceLabels
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxLabels'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLabels'
          description: Labels have been successfully replaced
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Replace sandbox labels
      tags:
        - sandbox
  /sandbox/{sandboxId}/backup:
    post:
      operationId: createBackup
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
          description: Sandbox backup has been initiated
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create sandbox backup
      tags:
        - sandbox
  /sandbox/{sandboxId}/public/{isPublic}:
    post:
      operationId: updatePublicStatus
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Public status to set
          explode: false
          in: path
          name: isPublic
          required: true
          schema:
            type: boolean
          style: simple
      responses:
        '201':
          description: ''
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Update public status
      tags:
        - sandbox
  /sandbox/{sandboxId}/autostop/{interval}:
    post:
      operationId: setAutostopInterval
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Auto-stop interval in minutes (0 to disable)
          explode: false
          in: path
          name: interval
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          description: Auto-stop interval has been set
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Set sandbox auto-stop interval
      tags:
        - sandbox
  /sandbox/{sandboxId}/autoarchive/{interval}:
    post:
      operationId: setAutoArchiveInterval
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Auto-archive interval in minutes (0 means the maximum interval
            will be used)
          explode: false
          in: path
          name: interval
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          description: Auto-archive interval has been set
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Set sandbox auto-archive interval
      tags:
        - sandbox
...
  /sandbox/{sandboxId}/ports/{port}/preview-url:
    get:
      operationId: getPortPreviewUrl
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Port number to get preview URL for
          explode: false
          in: path
          name: port
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortPreviewUrl'
          description: Preview URL for the specified port
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get preview URL for a sandbox port
      tags:
        - sandbox
  /sandbox/{sandboxId}/build-logs:
    get:
      operationId: getBuildLogs
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Whether to follow the logs stream
          explode: true
          in: query
          name: follow
          required: false
          schema:
            type: boolean
          style: form
      responses:
        '200':
          description: Build logs stream
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get build logs
      tags:
        - sandbox
...
  /workspace/{workspaceId}:
    delete:
      deprecated: true
      operationId: deleteWorkspace_deprecated
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
        - explode: true
          in: query
          name: force
          required: true
          schema:
            type: boolean
          style: form
      responses:
        '200':
          description: Workspace has been deleted
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: '[DEPRECATED] Delete workspace'
      tags:
        - workspace
    get:
      deprecated: true
      operationId: getWorkspace_deprecated
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Include verbose output
          explode: true
          in: query
          name: verbose
          required: false
          schema:
            type: boolean
          style: form
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: Workspace details
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: '[DEPRECATED] Get workspace details'
      tags:
        - workspace
  /workspace/{workspaceId}/start:
    post:
      deprecated: true
      operationId: startWorkspace_deprecated
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Workspace has been started
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: '[DEPRECATED] Start workspace'
      tags:
        - workspace
  /workspace/{workspaceId}/stop:
    post:
      deprecated: true
      operationId: stopWorkspace_deprecated
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Workspace has been stopped
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: '[DEPRECATED] Stop workspace'
      tags:
        - workspace
  /workspace/{workspaceId}/labels:
    put:
      deprecated: true
      operationId: replaceLabelsWorkspace_deprecated
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxLabels'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLabels'
          description: Labels have been successfully replaced
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: '[DEPRECATED] Replace workspace labels'
      tags:
        - workspace
...
  /workspace/{workspaceId}/archive:
    post:
      deprecated: true
      operationId: archiveWorkspace_deprecated
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Workspace has been archived
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: '[DEPRECATED] Archive workspace'
      tags:
        - workspace
...
components:
  schemas:
...
    SandboxLabels:
      example:
        labels:
          environment: dev
          team: backend
      properties:
        labels:
          additionalProperties:
            type: string
          description: Key-value pairs of labels
          example:
            environment: dev
            team: backend
          type: object
      required:
        - labels
      type: object
    PortPreviewUrl:
      example:
        url: https://123456-mysandbox.runner.com
        legacyProxyUrl: https://3000-mysandbox.runner.com
        token: ul67qtv-jl6wb9z5o3eii-ljqt9qed6l
      properties:
        url:
          description: Preview url
          example: https://123456-mysandbox.runner.com
          type: string
        token:
          description: Access token
          example: ul67qtv-jl6wb9z5o3eii-ljqt9qed6l
          type: string
        legacyProxyUrl:
          description: Legacy preview url using runner domain
          example: https://3000-mysandbox.runner.com
          type: string
      required:
        - token
        - url
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect